Add a check for empty handle in V8Proxy::GetSourceLineNumber()
If an exception occours in some of the code used to locate the line number of the source on the top stack frame an empty handle will be returned and using Int32Value on it will cause a crash. As of now I have not been able to find a case where an exception to happens though.

BUG=7912
Review URL: http://codereview.chromium.org/28024

git-svn-id: http://src.chromium.org/svn/trunk/src@10253 4ff67af0-8c30-449e-8e8b-ad334ec8d88c