Fixing crash in StartFinding (2nd attempt).
The first attempt to fix this was unsuccessful and crashes showed
up again during the automated ui testing. I still can't reproduce
locally, but when I look again at the crash dump I think that the
fact that it looked like the FindBarView was destroyed is misleading
and the real issue is the same as with the crash in StopFinding
(that web_contents has been destroyed moments before we receive the
request).

I am therefore removing the code I added earlier and replacing it
with a null check for web_contents.

TEST=Covered by automated ui tests.
BUG=8048

Review URL: http://codereview.chromium.org/27219

git-svn-id: http://src.chromium.org/svn/trunk/src@10489 4ff67af0-8c30-449e-8e8b-ad334ec8d88c