Make AutocompletePopupView a virtual interface.
The AutocompletePopupModel needs to be able to communicate with the view.  Since the view implementation will be platform specific, we abstract our a virtual interface that the different platform view implementations must all implement.  This allows the Model and View to communicate, while hiding the platform differences.

This change makes the previous Windows AutocompletePopupView into AutocompletePopupViewWin, a subclass of AutocompletePopupView.

Review URL: http://codereview.chromium.org/28258

git-svn-id: http://src.chromium.org/svn/trunk/src@10607 4ff67af0-8c30-449e-8e8b-ad334ec8d88c