Rebaseline form-element-geometry-expected.
There are basically two sets of changes from the last baseline, which was done by Finnur on Jan. 26.  One is a fix to SELECT scrollbars that looks like it's strictly more correct.  The other is the removal of a [border bgcolor="#000000"] on the BRs inside some SELECTs, which I'm less certain of, but which isn't present upstream and wasn't present in our baseline before the last set of changes to it.
Review URL: http://codereview.chromium.org/39268

git-svn-id: http://src.chromium.org/svn/trunk/src@11153 4ff67af0-8c30-449e-8e8b-ad334ec8d88c