Fixes the following layout tests.
1. LayoutTests\plugins\netscape-plugin-setwindow-size.html
2. LayoutTests\plugins\destroy-stream-twice.html.

The fix for the first test was to get our NPP_SetWindow implementation in the layout test plugin on par with webkit's implementation. I also added a fflush(stdout) to make sure that the printf shows up. If we don't do this the test fails often.

The fix for the second test is to get our NPP_DestroyStream implementation in the layout test plugin on par with webkit's implementation. With the latest webkit update, there is a destroyNullStream test which has been added in this layout test. The layout test plugin had to be updated with this new method on the same lines as its counterpart in  webkit.


Review URL: http://codereview.chromium.org/39242

git-svn-id: http://src.chromium.org/svn/trunk/src@11157 4ff67af0-8c30-449e-8e8b-ad334ec8d88c