Relanding a backed-out CL:
The navigation entry's state would not be returned for navigation entries of
error pages (such as 404).
This was causing the following problems when navigating back to a 404 page in
test shell:
- a new entry to be created instead of going back to the existing entry
- the navTarget property not to be set on the HistoryItem

That was causing the LayoutTests/http/tests/navigation/error404-goback.html
failure.
It seems safe to return the navigation entry's state for an error page.

BUG=None
TBR=darin
Review URL: http://codereview.chromium.org/43083

git-svn-id: http://src.chromium.org/svn/trunk/src@11453 4ff67af0-8c30-449e-8e8b-ad334ec8d88c