Relanding a backed-out CL that was thought to cause reliability test failure but that is not the culprit after-all:
The navigation entry's state would not be returned for navigation entries of
error pages (such as 404).
This was causing the following problems when navigating back to a 404 page in
test shell:
- a new entry to be created instead of going back to the existing entry
- the navTarget property not to be set on the HistoryItem

That was causing the LayoutTests/http/tests/navigation/error404-goback.html
failure.
It seems safe to return the navigation entry's state for an error page.

BUG=None
TBR=jar
Review URL: http://codereview.chromium.org/43100

git-svn-id: http://src.chromium.org/svn/trunk/src@11487 4ff67af0-8c30-449e-8e8b-ad334ec8d88c