Fix SessionRestoreUITest.TwoTabsSecondSelected in single process mode.
This is admitedly a little hacky, but I don't care as much since it's single-process mode and I don't think there's another way to solve it.  The problem is that SiteInstance::GetProcess should always reuse an existing renderer "process", but GetExistingProcessHost() won't return the in-process renderer until RenderProcessHost::Init is called, which is after the second GetProcess call.  So since we generate the in-process 'pid', I do it in the constructor instead of in Init.

Review URL: http://codereview.chromium.org/42144

git-svn-id: http://src.chromium.org/svn/trunk/src@11571 4ff67af0-8c30-449e-8e8b-ad334ec8d88c