Use anonymous shared memory for VisitedLinks.
We already pass a handle to the shared memory via IPC but, when we
create it, we create a named segment. These segments are not garbage
collected on POSIX.

Since nobody actually uses the name of the segment, use an anonymous
segment instead and remove the functions.

Review URL: http://codereview.chromium.org/48087


git-svn-id: http://src.chromium.org/svn/trunk/src@11900 4ff67af0-8c30-449e-8e8b-ad334ec8d88c