Fixing a few minor issues found during static analysis of the Chromium codebase.
find_bar_controller.cc:

Unused local variable created.

hwnd_view.cc:

The local definition of parent overshadows param passed in.

google_update.cc:

Uninitialized member variable.

webframe_impl.cc:

The if was overly constrained. At the top of the function there is an if (reset) { ... return; } so we don't need to check for !reset in the next if statement. This functionality is covered by a ui_test, which passes with this change.

Review URL: http://codereview.chromium.org/42321

git-svn-id: http://src.chromium.org/svn/trunk/src@11964 4ff67af0-8c30-449e-8e8b-ad334ec8d88c