Added a debug method AssertAcquired() that compiles to nothing in non-debug builds.  In debug builds, the method will DCHECK if the current thread does not hold the lock.
I also removed a class that was not being used (AutoLockImpl) which seems like someone just copied the Lock.h file and blindly added
an Impl for the AutoLock class.

Thre are places in my code where I want to use the AutoUnlock class.  I can't see why it was restricted to use by the condition variable
class only, so I just made it into a regular class with no restrictions.

I noticed that JamesR posted a CL about a month ago that made AutoUnlock a public class, but it was never committed, so I added him to this CL for review too.
Review URL: http://codereview.chromium.org/48109

git-svn-id: http://src.chromium.org/svn/trunk/src@12037 4ff67af0-8c30-449e-8e8b-ad334ec8d88c