ScopedVariant implementation.
A class for automatically freeing a COM VARIANT at the
end of a scope.  Additionally provides a few functions to
make the encapsulated VARIANT easier to use.
Instead of inheriting from VARIANT, I took the containment
approach in order to have more control over the usage of the
variant and guard against memory leaks.


Review URL: http://codereview.chromium.org/46059

git-svn-id: http://src.chromium.org/svn/trunk/src@12081 4ff67af0-8c30-449e-8e8b-ad334ec8d88c