Linux: back off build to only require P3
The current linux version of chromium will instantly crash with SIGILL
on an Athlon/AthlonXP/AthlonMP class processor.  It will presumably do
the same thing on the pentium3.  This is because it's built to use
SSE2 instructions to implement floating point.

We cannot build for i686 normally because we can't have the floating
point results be different between debug/opt builds. Without forcing
it to use SSE that happens (and it has a lot of noticeable effects on
our tests).

For builds which users actually use, I expect the distros will back
off the compiler flags because people won't be using those builds for
development.

However, I'm landing this patch because there's little point in
causing problems for SSE1 only chips because it will only make 3rd
party contributions harder. Thanks.

The only downside is minutely slower double precision floating point.

(Patch from Elladan)


git-svn-id: http://src.chromium.org/svn/trunk/src@12389 4ff67af0-8c30-449e-8e8b-ad334ec8d88c