Prevent data read errors from crashing the browser.
In ResourceDispatcherHost::OnResponseStarted(), if Read() causes an error, it calls OnResponseCompleted(). For a download, this hits the NOTREACHED() in DownloadThrottlingResourceHandler::OnResponseCompleted().

This patch checks for the error status, and returns false, which causes the download to finish, which saves an empty file of size 0.

BUG=8170
Review URL: http://codereview.chromium.org/42573

git-svn-id: http://src.chromium.org/svn/trunk/src@12410 4ff67af0-8c30-449e-8e8b-ad334ec8d88c