Chrome changes to support cached form submissions.
The solution is to add a user-defined identifier to UploadData.
If that identifier is set, and if the request method is POST,
then HttpCache will enable caching for the response.  (The cache
key will be a composition of the identifier and the URL.)  A
subsequent POST request to the same URL with the same identifier
will "hit" the previously generated cache entry.  Reuse from the
cache is subject to all of the standard rules.

For reference, here are the corresponding net changes:
http://codereview.chromium.org/52028

Here are the corresponding WebKit changes:
http://trac.webkit.org/changeset/41919

BUG=2636
R=sky
Review URL: http://codereview.chromium.org/52040

git-svn-id: http://src.chromium.org/svn/trunk/src@12485 4ff67af0-8c30-449e-8e8b-ad334ec8d88c