Changed the behavior of the form autofill popup. There are 3 ways to bring up the form autofill: 1 - start typing some text in a form text field 2 - press key down/up when the focus is on a form text field 3 - click an already focused form text field
In all these cases, our current behavior is to show the form autofill only if
the caret is at the end of the text field.

Other browsers' behavior:
IE shows the autofill popup in all 3 cases regardless of the caret position
FF shows the autofill popup regardless of the caret position for case 2 and 3
but not 1.
Safari never shows an autofill popup for cases 2 and 3, it shows the autofill
popup for case 1 only if the caret is at the end.

This CL changes our behavior to be like FF.

Also this CL now lets the default processing of key up/down events happen in the
case where we are showing an autofill. We were preventing default handling as it
moves the caret. Since were showing the autofill in a posted task and we would
check that the caret was at the end, this would prevent the autofill from
showing.
Now that we don't enforce the caret at the end condition with key up/down
presses, we can let the default handling move the caret.


BUG=6437
TEST=See description of 3 scenarios above. Make sure our behavior matches FF's.
Also test that password autocomplete still works.


Review URL: http://codereview.chromium.org/45067

git-svn-id: http://src.chromium.org/svn/trunk/src@12681 4ff67af0-8c30-449e-8e8b-ad334ec8d88c