The URL in the bug triggers an SSL interstitial for a redirect that redirects to a bad SSL page.
This happens because:
- when proceeding on an interstitial, we wait for the navigation to commit before hiding the current interstitial
- when showing an interstitial, if an interstitial is already showing we call DontProceed on it

In the case of this bug, the interstitial has already been approved when the showing of the 2nd interstitial calls DontProceed on it, causing bad things to happen.

The fix is easy: we shouldn't call DontProceed on an interstitial already "Proceeded".


BUG=9286
TEST=See bug.
Review URL: http://codereview.chromium.org/56019

git-svn-id: http://src.chromium.org/svn/trunk/src@12727 4ff67af0-8c30-449e-8e8b-ad334ec8d88c