TBR: Skip ExtensionContentScriptInjectTest.Simple in purify.
I give up. Not sure why the filters aren't working, but I
suspect that even if I get them working, they will be becoming
invalid all the time anyway as the codebase changes.

Too bad, because the information we got from purify was useful
(it found a couple memory leaks in other parts of Chrome) but
I really need to get back to work. And I think that having this
unit test running, even without purify, is more valuable than
not.

I created a bug on this and assigned it to myself. Hopefully as
more people do high-level tests like this, the bugs will
slowly be worked out. Otherwise, I'll eventualy get back to it.

Review URL: http://codereview.chromium.org/62173

git-svn-id: http://src.chromium.org/svn/trunk/src@13403 4ff67af0-8c30-449e-8e8b-ad334ec8d88c