Removing --enable-video and instead relying on having DLLs present.
I measured a ~10ms startup regression when the DLLs are not present and a ~20ms
regression when the DLLs are present and loaded in render_process.cc.

Not sure how else to tackle this problem since we need these DLLs loaded before
the sandbox is engaged.

Review URL: http://codereview.chromium.org/63101

git-svn-id: http://src.chromium.org/svn/trunk/src@13542 4ff67af0-8c30-449e-8e8b-ad334ec8d88c