Reduce the false positive rate for SafeBrowsing gethash requests.
This CL increases the memory consumption of the bloom filter used
by the SafeBrowsing system in order to decrease the number of false
positive gethash requests (gethash requests that result in an empty
or 204 response from the servers).

The filter size in bytes is calculated as:
  number_of_add_prefixes * bits_per_prefix / 8

From analysis of our histograms, users have between 250k - 330k add
prefixes. 'bits_per_prefix' is hard coded to 25, which means that we
expect the typical bloom filter to be between 760-1000 kB, compared
to the current value of approximately 450 kB.

We add histograms to track the filter size, as well as the number of
gethash requests that return empty results and non-empty results.

BUG=10584 (http://crbug.com/10584)

Review URL: http://codereview.chromium.org/67243

git-svn-id: http://src.chromium.org/svn/trunk/src@13958 4ff67af0-8c30-449e-8e8b-ad334ec8d88c