Fixes FFmpegDemuxerTest.ReadAndSeek test flakiness by using a WaitableEvent.
Turns out it was the test code that was flaky and not a threading bug in our actual code.  Under heavy load it was possible for a thread holding onto a reference to get time sliced long enough that the unit test continued executing until it asserted that the object was deleted when in fact it was not.

BUG=10653

Review URL: http://codereview.chromium.org/88010

git-svn-id: http://src.chromium.org/svn/trunk/src@14113 4ff67af0-8c30-449e-8e8b-ad334ec8d88c