Print backtraces on FATAL log messages in debug mode.
This provides basic support for looking up backtrace information on GNU libc systems and in Windows.  The code is only enabled for FATAL log messages in debug mode.  In a release build, it is unlikely that symbols will be available making the backtrace less useful.

Review URL: http://codereview.chromium.org/62140

git-svn-id: http://src.chromium.org/svn/trunk/src@14391 4ff67af0-8c30-449e-8e8b-ad334ec8d88c