linux: make --single-process mean in-process renderer.
Since we can't have in-process plugins on Linux, let's make
single-process do as much of a single process as possible.

The alternative is detecting and erroring out on --single-process, then
adding a new flag and changing all of the other code that currently
examines --single-process to examine both the new flag and
--single-process.  That seems like a lot of extra code and source of
potential error.

Note that we print out a warning indicating this on --single-process
in the LINUX case.

Review URL: http://codereview.chromium.org/100053

git-svn-id: http://src.chromium.org/svn/trunk/src@14645 4ff67af0-8c30-449e-8e8b-ad334ec8d88c