Don't start the throbber when loading javascript URLs. This fixes a regression from my "start the throbber sooner" patch that made JS URLs keep throbbing forever.
This patch just adds a check to see if the URL is JS, and doesn't synthesize
the load start message. It seems kind of like a hcak, but I think it's the
right thing. As I argued in the last patch (the comment for the block I
modified here), we really need the throbbing status to be synchronous. This
means we need to basically duplicate the logic for when WebKit will do the
throbber in this function.

BUG=11422
Review URL: http://codereview.chromium.org/105009

git-svn-id: http://src.chromium.org/svn/trunk/src@15280 4ff67af0-8c30-449e-8e8b-ad334ec8d88c