Convert autocomplete to use a custom matrix and button cell.
On the user-visible side this change enables mouse hover (the item
under the mouse is highlighted), and lays out the text omnibox v2
style.  Only hover really needed the control change, but I have
another change queued up to style the popup text.  It uses different
colors and shades, and really looked horrible with the NSTableView
highlighting.  [Which is why this part of the CL is coming first.]

TEST=Omnibox popup works and shows highlighting under the mouse in addition to the keyboard-selected item.

Review URL: http://codereview.chromium.org/99310

git-svn-id: http://src.chromium.org/svn/trunk/src@15321 4ff67af0-8c30-449e-8e8b-ad334ec8d88c