Make background pages work more nicely with the tabs API.
Changed ExtensionHost so that it returns NULL if it doesn't have
a browser, instead of getting the last active one. The problem
is that if ExtensionHost returns the last active browser at the
time it's constructed, it might be garbage by the time it is
used.

Changed tab functions to use more consistent logic to determine
the 'current' browser to operate on.

Review URL: http://codereview.chromium.org/115071

git-svn-id: http://src.chromium.org/svn/trunk/src@15587 4ff67af0-8c30-449e-8e8b-ad334ec8d88c