Fix for b/1837411 -- Interstitial title lines bleed together.
The line height for titles in the warning interstitials is too small. 
In English, a second line starts to abut the line above it if the 
browser is narrow.

In languages with ligatures above the line, a second line bleeds 
into the first, impairing readability. This is compounded in languages
like Tamil and Marathi, where the title of the interstitial warning 
is quite long and likely to break on two lines.

My solution is to set the line height to 140% in the .title.

TEST= Navigate to ianfette.org with Chrome's lang set to ta or mr. The title 
lines in the anti-malware blocking page should not bleed together.
BUG= 1837411
Review URL: http://codereview.chromium.org/113140

git-svn-id: http://src.chromium.org/svn/trunk/src@15667 4ff67af0-8c30-449e-8e8b-ad334ec8d88c