New baselines to account for http://trac.webkit.org/changeset/43355
Note: that change only required a small fraction of tests to be rebaselined for
PLATFORM(MAC).  However, due to the change made in themeWin.css, a lot more
tests are impacted for PLATFORM(CHROMIUM) on Windows.

I expect that these will also need to be re-baselined for Linux separately
since the change involves the rendering of form controls.

R=dglazkov

Review URL: http://codereview.chromium.org/113170

git-svn-id: http://src.chromium.org/svn/trunk/src@15697 4ff67af0-8c30-449e-8e8b-ad334ec8d88c