Decrease the SafeBrowsing false positive rate.
This CL replaces the weak hashing functions of the bloom
filter for a more thorough mixing function. The bloom
filter is modified to use and store random keys in the
mixing function. This addresses the need for randomizing
the contents of each user's bloom filter to prevent a
false positive for a popular site causing a global spike
in hash requests.

BUG=10584 (http://crbug.com/10584)
BUG=10634 (http://crbug.com/10634)

Review URL: http://codereview.chromium.org/113064

git-svn-id: http://src.chromium.org/svn/trunk/src@15703 4ff67af0-8c30-449e-8e8b-ad334ec8d88c