Style the omnibox results for v2.
The style is based on the gtk code (autocomplete_popup_view_gtk.cc), with some local tweaks.  I broke this off from the rest of the review so I could experiment with making it nail omnibox-v2, but eventually realized that I don't have enough data for that.  This at least makes things look nicer, and I believe it provides all of the function that will be needed (the remainder is precisely defining fonts and colors and possibly minor highlighting variations).

Summary: I want something concrete so that complaints about v2 wrongness can be specific.

Review URL: http://codereview.chromium.org/99312

git-svn-id: http://src.chromium.org/svn/trunk/src@15777 4ff67af0-8c30-449e-8e8b-ad334ec8d88c