Implemented Pipeline seek callbacks and a big refactor of PipelineImplTest.
Now that I'm wiser writing tests, I'm trying to get rid of InitializationHelper and rely on simplier test fixtures classes.  I also introduced MockFilterFactory which is capable of creating all the mock filters and is able to return instantiated instances of filters to permit testing for expectations (refer to PipelineImplTest.Seek for an example).

Review URL: http://codereview.chromium.org/115094

git-svn-id: http://src.chromium.org/svn/trunk/src@15829 4ff67af0-8c30-449e-8e8b-ad334ec8d88c