Factor out a BrowserFrame interface, and move the existing Windows-based code to a new class: BrowserFrameWin.
The major other change I had to make was the way that the TabStrip is set in
the BrowserRootView. Now there is a method on the frame that the view can
call when it creats the tabstrip that will set the tabstrip on the root view.
Review URL: http://codereview.chromium.org/113511

git-svn-id: http://src.chromium.org/svn/trunk/src@16373 4ff67af0-8c30-449e-8e8b-ad334ec8d88c