Use av_rescale_q() for converting FFmpeg timestamps to base::TimeDelta.
Previously we were using integer math to convert to microseconds, but depending on the frame rate and packet size we could introduce enough error that could accumulate and introduce audio/video synchronization drift.  av_rescale_q() is a simple function but is designed to minimize error as much as possible.

Review URL: http://codereview.chromium.org/113598

git-svn-id: http://src.chromium.org/svn/trunk/src@16453 4ff67af0-8c30-449e-8e8b-ad334ec8d88c