Attempting to make the Gmail checker more robust in the face of errors.
The heart of this is that I wrapped up the request to the feed in a timeout and a bunch more error handling and abstracted all that into one function so I could be sure all the errors were getting handled the same way.

Also general cleaning while I was in there.

BUG=11969
TEST=Install Gmail Checker, go offline, go back online. It should start showing message counts again when you go back online.

Review URL: http://codereview.chromium.org/115584

git-svn-id: http://src.chromium.org/svn/trunk/src@16522 4ff67af0-8c30-449e-8e8b-ad334ec8d88c