Properly dismiss cocoa popup menus.
The virtual key code for mac popup menus was not set properly
so HTML select menus, when dismissed by hitting escape or
clicking outside of the popup menu area, were not shut down
cleanly.

WebInputEventFactory requires the charactersIgnoringModifiers
field of the NSEvent to be set to VK_ESCAPE (0x1b) for the 
dimiss to work.

The key event type also needs to be a key down (NSKeyDown)
since PopupListBox::handleKeyEvent ignores the key up.

Review URL: http://codereview.chromium.org/115618

git-svn-id: http://src.chromium.org/svn/trunk/src@16659 4ff67af0-8c30-449e-8e8b-ad334ec8d88c