Introduce a buffer wait and timeout threshold for BufferedResourceLoader
The logic in the seek method of BufferedResourceLoader was
that it will wait until the data is downloaded if we know
that the desired seek position is within the bufferable range
of the buffer, which is 10MB for now. This number is found to
be too big and we should have another constant that governs
the threshold of bytes that we should wait for forward
seeking instead of starting a new request, this value is
essentially much smaller than the buffer size, we guess this
number to be 2MB.

Also introduced the timeout logic to BufferedResourceLoader,
so if BufferedResouceLoaderRead() takes more than a certain
amount of time it is declared timeout and a new request
should be started. The current timeout is set to 5s.

Review URL: http://codereview.chromium.org/115627

git-svn-id: http://src.chromium.org/svn/trunk/src@16793 4ff67af0-8c30-449e-8e8b-ad334ec8d88c