A quick fix for Issue 11863 (Take 2). This change is exactly the same change as <http://codereview.chromium.org/115353>. I would like to re-send this change because the PC that I created this original change has been dead and I have not been able to commit.
This issue is caused by a stupid mistake of mine that I forgot adding a WM_IME_ENDCOMPOSITION handler to update the find results when a composition is finished (or canceled). This change adds a handler for WM_IME_ENDCOMPOSITION messages to update the find results.

BUG=11863 "Korean find-in-page: Backspace all the way, Chrome will highlight all the characters that contain the first consonant"
TEST=Open a Korean page, type Control+F keys, Switch the input language to Korean, type a right-alt key, type a 'w' key, type a 'k' key, type a 'a' key, type a backspace key, and verify there are not any highlited characters. 
Review URL: http://codereview.chromium.org/113847

git-svn-id: http://src.chromium.org/svn/trunk/src@16934 4ff67af0-8c30-449e-8e8b-ad334ec8d88c