linux: simplify gfx::Font constructor logic and add CHECK()s.
Trying to track down a bug that a user is seeing.
The getter we were using to extract the value before was for
a debugger-compatible string.  Perhaps it was the quote-removal
code that was getting us before; if not, these CHECKs will track
it down.

BUG=12530

Review URL: http://codereview.chromium.org/115877

git-svn-id: http://src.chromium.org/svn/trunk/src@17118 4ff67af0-8c30-449e-8e8b-ad334ec8d88c