Python script for auto-generating stub/def files for a list of signatures.
This python script takes a list of C-style signatures and generates either a windows def file, or a C++ stub file that allows for binding of functions via dlopen/dlsym.  This is used to support delay loading in windows and simulate it on posix systems.

Review URL: http://codereview.chromium.org/113803

git-svn-id: http://src.chromium.org/svn/trunk/src@17534 4ff67af0-8c30-449e-8e8b-ad334ec8d88c