Don't update our text from toolbar if the toolbar's text doesn't visibly differ from the model's text.
In most cases, this doesn't come up, because focus is usually not in
the field (so there wouldn't be local changes), but when opening NTP
you can type between when the tab opens in the browser and when the
renderer finishes loading it.  In that case, the model and toolbar
both have "" as the url text, so this case is hit.

http://crbug.com/13109
TEST=Browse to www.google.com.  Command-t for new tab and type http as fast as you can.  Before this fix, you sometimes got "tp" or something of the sort.

Review URL: http://codereview.chromium.org/119125

git-svn-id: http://src.chromium.org/svn/trunk/src@17662 4ff67af0-8c30-449e-8e8b-ad334ec8d88c