Linux: Fake a mouse move event when the render view gains focus and the mouse is over it.
On windows, when the render view regains focus it receives a mouse move event with the coordinates of the pointer. We don't on linux, so we have to fake this. Without this, webkit doesn't know where the pointer is, so can have a wrong cursor or other graphical glitches.

TEST=bug is fixed
BUG=12879

Review URL: http://codereview.chromium.org/113987

git-svn-id: http://src.chromium.org/svn/trunk/src@17693 4ff67af0-8c30-449e-8e8b-ad334ec8d88c