Add this ability to install Extensions using preferences. Also known as: port the installation mechanism to other platforms.
We already have the ability to install extensions using a registry key. That works only on Windows so this new change adds the same but using preferences instead of the Registry. This will eventually allow us to pre-install certain extensions when we install Chrome.

BUG=12060
TEST=Covered by unit tests, but to test manually: close Chrome, open your Preferences file (in your profile) and add this (after substituting all <values> in elbow brackets):

   "extensions": {
      "settings": {
         "<your_extension_id_lowercased>": {
            "external_crx": "<path_to_crx>",
            "external_version": "<crx version>"
         }
      },
   },

... then start Chrome. Your extension should get installed. 
Review URL: http://codereview.chromium.org/119195

git-svn-id: http://src.chromium.org/svn/trunk/src@17777 4ff67af0-8c30-449e-8e8b-ad334ec8d88c