Fix an issue where themes would sporadically fail to install.
Trying to send decoded images over IPC didn't work too well.  Instead, we'll
write them to a file and have the browser slurp them in from there.  My first
instinct was to use SharedMemory, but that would require us to impose a limit
on the size of the decoded image data.

Also made sure that the undecoded images are deleted when we install.

BUG=13455
TEST=Try the repro steps in bug 13455 several times and make sure it works
every time.
Review URL: http://codereview.chromium.org/119255

git-svn-id: http://src.chromium.org/svn/trunk/src@17797 4ff67af0-8c30-449e-8e8b-ad334ec8d88c