Stats for nerds should bring up applicaiton browser window to foreground
The reason why it didn't bring up the stats in foreground is because it is an application window. Application (same goes for popup) windows cannot have tabs, it is just a single page. When the link is clicked, we need to get the last active browser so we could open that about:memory page. When you open a page in an application window, it will create a new window or find a window (normal) that is present, then open the url in that tab. So now we don't have the browser pointed to that tab, its still pointed to the application window.

By finding the last active tab again, we can ensure we have the correct tab.

BUG=11822 (http://crbug.com/11822)
TEST=tested if the stats appeared in foreground and it did.
Review URL: http://codereview.chromium.org/119293

git-svn-id: http://src.chromium.org/svn/trunk/src@17925 4ff67af0-8c30-449e-8e8b-ad334ec8d88c