Should provide checks while loading bookmarks
Instead of providing a separate check to validate bookmarks while loading, we could assume invalid bookmarks are invalid, hence we don't break the loading sequence. We keep looping since another bookmark might be valid.

If the JSON syntax is broken, it will exit gracefully because the parser can't parse the file. This patch is only good for people who are editing their bookmarks directly and missing a item in the bookmark node.

BUG=10709 (http://crbug.com/10709)

TEST=Deleted any line from the Bookmarks file within the bookmark children node. Loaded up Chromium and it loaded the error free bookmarks.

Review URL: http://codereview.chromium.org/118351

git-svn-id: http://src.chromium.org/svn/trunk/src@17960 4ff67af0-8c30-449e-8e8b-ad334ec8d88c