First cut at tall titlebar and custom window frame.
This implements the tall titlebar with the XP frame buttons, which should
work if your window manager supports them.

The titlebar background is drawn by the containing hbox.  It would be nice
if the tab strip could be made transparent so it wouldn't need to paint
a background itself, but I didn't know how to do that.

I still need to implement the frame resizing, frame coloring and move the
context menu from the tab strip to the titlebar.

BUG=13430

Review URL: http://codereview.chromium.org/119371


git-svn-id: http://src.chromium.org/svn/trunk/src@18107 4ff67af0-8c30-449e-8e8b-ad334ec8d88c