Fix HWND for Importer Dialog
Instead of grabbing the HWND from the browser, we can grab it from the current view since each view has a Widget who is associated with a HWND (NativeView). By doing that, we can assure the parent hwnd is that. Before, it was retrieving the hwnd from Browser* which caused this bug.

BUG=13154 (http://crbug.com/13154)

TEST=Launched Options>Importer>Import and it brought up the Options dialog.

Review URL: http://codereview.chromium.org/119408

git-svn-id: http://src.chromium.org/svn/trunk/src@18118 4ff67af0-8c30-449e-8e8b-ad334ec8d88c