Extract form related classes from the guts of WebFrameImpl.
Instead of having WebFrameImpl generate SearchableFormData, PasswordForm, and AutofillForm classes, allow the embedder (RenderView) to do so.

This is done to help minimize the dependencies WebFrameImpl has on other code, which will make it easier to move WebFrame and WebDataSource into the WebKit API.

Most significant change:  Now, RenderView always sets a NavigationState on WebDataSource instances.  We used to only do so for browser initiated navigations.  This is done so that we can store things like SearchableFormData and friends on the NavigationState.

To facilitate this change, it was necessary to add a way through the WebKit API to refer to a HTMLFormElement.  This CL introduces WebForm, which is like a RefPtr<HTMLFormElement>, so you can just copy a WebForm around by value and the right thing happens.

Some of the other changes are about moving more things into the webkit_glue namespace.  On hindsight, I probably should have done that as a separate CL.

BUG=10041
TEST=none
R=brettw
Review URL: http://codereview.chromium.org/126083

git-svn-id: http://src.chromium.org/svn/trunk/src@18395 4ff67af0-8c30-449e-8e8b-ad334ec8d88c