Speculative fix for a chromebot crash since I can't reproduce it locally.
When shutting down the browser with a download shelf, we first
release the shelf (see BUG=13681 for details) but we are still
part of the view hierarchy. A little later in the shutdown, the
shelf's parent (BrowserView) will dereference the already deleted
shelf through its View base class, which I believe is the source
of the Chromebot crash.

BUG=14171 (http://crbug.com/14171)
TEST=Chromebot should not crash

Review URL: http://codereview.chromium.org/126228

git-svn-id: http://src.chromium.org/svn/trunk/src@18532 4ff67af0-8c30-449e-8e8b-ad334ec8d88c