Properly cleanup state after synchronous DNS resolution
Pre-resoultion system was not cleaning up its state when the resolver
returned asynchronously.  This regression probably landed when we shifted
to using an internal resolver class, rather than running on worker threads.
The impact was made detectable by the recent introudction of a micro-cache,
which often returns with a synchronous result.

BUG=14212

r=eroman
Review URL: http://codereview.chromium.org/125209

git-svn-id: http://src.chromium.org/svn/trunk/src@18537 4ff67af0-8c30-449e-8e8b-ad334ec8d88c