Re-land another subset of r18520.
This particular subset should be a no-op: It adds an unused interface method (OnResolutionCancelled), and an unused test (CancelObserver), and lastly renames Cancel --> MarkAsCanceled().
The code in "dns_global.cc" is also unused since it is unreachable.

The original code review that this belongs to is:<http://codereview.chromium.org/125171>. (BUG=14138)

The original was backed out because of a valgrind linux error (BUG=14218), so I am now re-landing in smaller chunks to identify where the problem resides.
Review URL: http://codereview.chromium.org/126253

git-svn-id: http://src.chromium.org/svn/trunk/src@18581 4ff67af0-8c30-449e-8e8b-ad334ec8d88c