Allow the RenderWidgetHostViewGtk to be sized even if it isn't a popup.
browser/views/tab_contents/tab_contents_view_gtk.cc implements sizing in a similar fashion to windows - i.e. when the GtkWidget is resized, it tells the RenderViewHost that it was sized, which causes the RVH to resize the view. A check for if (parent_) in the Gtk RenderWidgetHostView impl was preventing this sizing from occurring.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/125275

git-svn-id: http://src.chromium.org/svn/trunk/src@18670 4ff67af0-8c30-449e-8e8b-ad334ec8d88c