When we upgrade, discard all uploadable stats, including old logs
We already discard all stability info when we do an upgrade to a new
version, as there is less concern about stats from old versions. This
change just adds the list of "old logs" (gathered in previous sessions,
presumably with an old version) to the list of data that is discarded.

BUG=14657

r=evanm
Review URL: http://codereview.chromium.org/131089

git-svn-id: http://src.chromium.org/svn/trunk/src@18794 4ff67af0-8c30-449e-8e8b-ad334ec8d88c