Fix: New window appears to the left/up if the taskbar is on left/top.
Currently, the Windows taskbar (put differently, work area) is not considered
in getting the last window bounds (while it is for the saved window bounds).
This causes new windows (CTRL-N) appear to the left/up of the existing window
instead of right/down, if the taskbar is on left or top.

TESTED=gcl try, manually

TEST=Place the taskbar on left or top of the screen and create new Chrome windows by CTRL-N and observe.

BUG=14131

Review URL: http://codereview.chromium.org/125179

git-svn-id: http://src.chromium.org/svn/trunk/src@18795 4ff67af0-8c30-449e-8e8b-ad334ec8d88c