Enable DownloadRequestManager for Mac.
DownloadRequestManager is the object that manages the number of downloads
that can occur in a given tab. It throttles non-user initiated downloads
to prevent carpet-bombing download attacks.

Enabling this object for Mac fixes the issue where we can only download
one file in a given tab.

BUG=None (mac work).
TEST=Can now download more than one item in a tab.

Review URL: http://codereview.chromium.org/132039

git-svn-id: http://src.chromium.org/svn/trunk/src@18840 4ff67af0-8c30-449e-8e8b-ad334ec8d88c