Minor no-op change to ASN.1 encoding of private key.
The logic for when to prepend a null byte to integers was
reversed. Fortunately, this had no impact on anything, because:

a) We ignore the null byte on the parse side because we know the
integers we're interested in are always unsigned.

b) It appears openssl does the same thing (I tried it both ways
and it always came up with the same private key data).

Still, better to be correct.

Also, re-enable previously disabled unit tests.

BUG=14640

Review URL: http://codereview.chromium.org/131095

git-svn-id: http://src.chromium.org/svn/trunk/src@18855 4ff67af0-8c30-449e-8e8b-ad334ec8d88c