Handle proxy corruption including nesting of compressino algorithms
Vodaphone (UK) Mobile Broadband appears to be providing an extra level
of gzip compression, at the same time as it discards the content encoding
"sdch,gzip" and replaces it with "gzip."

The previous approach for handling a missing "sdch" content encoding 
statement from a server that *probably* was trying to send encoded 
content, was to replace the (probably) garbage content encoding with 
tentative decodings for both gzip and sdch.  That approach was not 
sufficient for the resulting double-gzip encoding that Vodaphone 
provided :-(.

This fix leaves the existing content encodings (such as Vodaphone's
solo "gzip"), and adds the tentative decodings (gzip plus sdch).  

Hence we translate received content encodings (when we know we requested
an SDCH response by including an applicable dictionary) as follows:

"sdch,gzip"  unchanged.  This is perfect.
""  ==> "Tent_sdch, Tent_gzip"
"gzip" ==> "Tent_sdch, Test_gzip, gzip"

Note that the TENTATIVE_gzip is a sniffing gzip, that can turn into an 
identity (no-op) filter if it can't find a gzip header.

By continuing to use a tentative SDCH, it is possible for the SDCH 
filter to make error recovery decision, and to record stats
on this case.

In addition, I've changed the error recovery approach in the SDCH 
filter to be ultra conservative.  If there is ANY unrecoverable 
error in a response to an SDCH request, then we will blacklist 
the host of the URL.  Recoverable errors are where the SDCH is fully
functional, and decodes its data using the dictionary.

The test cases added focus on the double gzip cases, as observed in
Vodaphone (UK) Mobile Broadband.

BUG=13606

r=huanr
Review URL: http://codereview.chromium.org/140037

git-svn-id: http://src.chromium.org/svn/trunk/src@18936 4ff67af0-8c30-449e-8e8b-ad334ec8d88c