GTK: Fix status bubble hiding during background tab events.
So the GTK version had only one status, setting the status to any URL the user hovered over. When the background tab is done with loading, it sends a SetStatus("") to the StatusBubble. This was what was tickling the bug.

So I'm copying the Windows implementation, which can have a URL and a status, with the URL's display overriding the status display. So when the background tab calls SetStatus(""), the URL should still be displayed.

http://crbug.com/14802

Review URL: http://codereview.chromium.org/147050

git-svn-id: http://src.chromium.org/svn/trunk/src@19077 4ff67af0-8c30-449e-8e8b-ad334ec8d88c