Fix issue where the closed window menu got cut off.
Also, fix issue where the recent items could not be executed using the
keyboard.

BUG=15104, 15111

TEST=Open a new window with 5+ tabs and close it. Then go to the new
new tab page and mouseover the item "5" tabs". It should show the whole
menu and you should also be able to tab to it and pressing enter when
focused should reopen the window.

Review URL: http://codereview.chromium.org/147054

git-svn-id: http://src.chromium.org/svn/trunk/src@19093 4ff67af0-8c30-449e-8e8b-ad334ec8d88c