implement WebSandboxSupport stub to prevent a null ptr deref
If FontChromiumWin's Font::drawGlyphs() fails for some reason, we
may attempt to reload the font using ChromiumBridge::ensureFontLoaded().
This requires a sandbox routine, SandboxSupport::ensureFontLoaded() to be
implemented, and in test_shell, it wasn't, causing a null ptr deref and
weird things to happen.

Note that there is no easy way to provoke this codepath once the other
bugs causing the crash in #10977 are fixed.

R=darin
BUG=http://crbug.org/10977
TEST=none


git-svn-id: http://src.chromium.org/svn/trunk/src@19200 4ff67af0-8c30-449e-8e8b-ad334ec8d88c