The Provider for External Extensions no longer uses the prefs file, but instead reads a preferences file (a json file called external_extensions.json).
This file specifies what external extensions we
installed (shipped with Chrome) and the extension
system takes care of installing them.

Having this in a separate file makes the upgrade
scenario easier, since we can during upgrade simply
blow away this file and replace it (and not have
to worry about merging/loosing user specified data).

The extension system will read the file, and figure
out whether any extension needs to be upgraded based
on the information provided.

Now, all external providers fetch data on the file
thread instead of reading and copying preferences on
the UI thread and posting them to the file thread.

BUG=http://crbug.com/14201
TEST=Covered by unit test.
Review URL: http://codereview.chromium.org/144018

git-svn-id: http://src.chromium.org/svn/trunk/src@19264 4ff67af0-8c30-449e-8e8b-ad334ec8d88c