Linux: paste on middle-click up.
There is some debate about when pasting should actually happen
   xterm: pastes on up.
   GTK: pastes on down.
   Firefox: pastes on up.
   Midori: couldn't paste at all with 0.1.2

There is something of a webcompat angle to this well, as highlighted by
crbug.com/14608. Pages can clear text boxes 'onclick' and, if we paste on
down then the text is pasted just before the onclick handler runs and
clears the text box.

http://codereview.chromium.org/149020
BUG=14608


git-svn-id: http://src.chromium.org/svn/trunk/src@19304 4ff67af0-8c30-449e-8e8b-ad334ec8d88c