linux plugins: eliminate GtkWidget* from plugin process
[retry with mac hopefully fixed this time]

GtkWidget* (and its wrapper, gfx::NativeView) only work within a
single process.  Plugins already work with a lower-level type that
works across processes -- an X Window id.  The parent of a plugin
is an HWND on windows, but it's an X Window id on X.  So we
introduce a new typedef wrapping that and push it through all the
places in the code that needs it.

Since we no longer have a GtkSocket in the WebPluginDelegateImpl,
we need to do a bit more work in the WebViewDelegate (aka the
browser process in the multiproc world).

We also need the plugin host (the browser) to track the
GtkSockets that are hosting the plugin, as well as destroy them
when necessary.  To do this we require the plugin owner to
grab the plug-removed signal rather than putting its handler in
GtkPluginContainer; this is the way it worked before I'd refactored
into GtkPluginContainer so it shouldn't be so bad.

This change still only works in test_shell, but the refactoring
should translate to the multiprocess case pretty easily.

Review URL: http://codereview.chromium.org/146009

git-svn-id: http://src.chromium.org/svn/trunk/src@19387 4ff67af0-8c30-449e-8e8b-ad334ec8d88c