Refactor extension bindings to share code, avoid exposing hidden variables globally, and avoid using the DOM load/unload events. - moved callback handling into event_bindings.js (ports will use it). - added chromeHidden, a V8 hidden value, to keep all internal variables that   need to be accessible to native code. - changed context registration to occur always at extension load, instead of   DOM load. - added an internal unload event that doesn't disable SuddenTermination.
This is a rework of my earlier CL http://codereview.chromium.org/125280 which was reverted because of a perf regression.  I believe the perf problem was caused by the call into javascript I did on page load to handle context registration - this CL avoids that.

Review URL: http://codereview.chromium.org/147033


git-svn-id: http://src.chromium.org/svn/trunk/src@19634 4ff67af0-8c30-449e-8e8b-ad334ec8d88c