Adds tab pinning to TabStripModel. This is just the model side of things (and tests), no UI yet.
The model enforces that all pinned tabs occur at the beginning of the
model. For example, if there are no pinned tabs and you pin the 10th
tab, it is moved to the front of the tab strip. Similarly inserting a
tab before the last pinned tab implicitly pins it. Moving a pinned tab
beyond the set of pinned tabs implicitly unpins it.

I'll file a bug on this when the site allows me to.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/155228

git-svn-id: http://src.chromium.org/svn/trunk/src@20181 4ff67af0-8c30-449e-8e8b-ad334ec8d88c