plugins: use OnChannelError to detect when the channel goes away
Previously we used a special Watcher object to watch the process,
but that is not portable to POSIX and this is simpler anyway.
With this change, I now see ~PluginProcessHost() running when a
plugin crashes.

If you recall all the way back to
  http://codereview.chromium.org/16814
we did a similar thing to the renderer host.

Review URL: http://codereview.chromium.org/155331

git-svn-id: http://src.chromium.org/svn/trunk/src@20414 4ff67af0-8c30-449e-8e8b-ad334ec8d88c