More media::PipelineImpl cleanup, this time focusing on not taking down the render process.
Specifically, if the pipeline hasn't started we should fail gracefully and not DCHECK.  Volume and playback rate are now allowed to be set if the pipeline isn't running, and filters will now receive a call to SetVolume() and SetPlaybackRate() with the initial values when the pipeline has fully initialized.

Added tests and expectations for all of this and ran valgrind to verify that we were indeed leaking memory (now fixed).

BUG=16009, 13902
TEST=media_unittests, layout tests

Review URL: http://codereview.chromium.org/149500


git-svn-id: http://src.chromium.org/svn/trunk/src@20455 4ff67af0-8c30-449e-8e8b-ad334ec8d88c