Lands http://codereview.chromium.org/149404 for cmasone:
Adding a protobuf-based metrics communication scheme between Chrome OS and
Chrome.  Essentially, Chrome is used as a metrics upload service, and this hacks
in a simple IPC mechanism for Chrome OS to use to communicate metrics info to
Chrome for upload.  We eventually want to do this using protobuffers over DBus,
but this protobuffers-over-X-properties mechanism will work for now.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/149459

git-svn-id: http://src.chromium.org/svn/trunk/src@20539 4ff67af0-8c30-449e-8e8b-ad334ec8d88c