Middle clicks on all fragment links open in new tab.
Middle clicking on an anchor tag that points to an
in-doc fragment should result in a new background
tab being opened.  Before this change, we were
overriding this behavior to cause a middle click to
behave like a left click.  This patch allows middle
clicks to function similar to the behavior seen in
Firefox and Safari.

No change in behavior should be seen by users for left
clicks on anchor tags that refer to in-doc fragments.

While testing, I see at times issues due to scrolling
within the tab to the correct fragment identifier or
start of the page.  This has already been identified in
issues such as 9319, so I'm ignoring that for now.

BUG=2913
TEST=Load Wikipedia, visit a page with many subheadings,
left click on a subheading, verify the current page
scrolls to the correct heading.  Click back, verify you
are taken back to your original position.  Middle click
on the same subheading, verify a new tab is loaded in
the background that is scrolled to the correct heading.

Review URL: http://codereview.chromium.org/155522

git-svn-id: http://src.chromium.org/svn/trunk/src@20674 4ff67af0-8c30-449e-8e8b-ad334ec8d88c