test_shell: set a title on the home page
This seems pointless, but for debugging purposes internally we
use the page title as an annotation on OS-level windows which helps
track them down when looking at a list of all windows.  For
reasons too complicated for me to care about this doesn't work when
a page lacks a title, hence this fix.

Review URL: http://codereview.chromium.org/149642

git-svn-id: http://src.chromium.org/svn/trunk/src@20691 4ff67af0-8c30-449e-8e8b-ad334ec8d88c