Fix: Linux file:// listings are sorted but not fully internationalized
Changed the code such that file names are converted to UTF-16 before
comparison. Ideally, the locale should be considered, but my conclusion is that
ignoring it is a reasonable tradeoff. Passing a locale-aware comparator
(collator) from src/chrome(or src/app?) to src/base requires too much effort, in
my opinion.

BUG=16179

TEST=Open a file:// URL for a directory that contains non-ascii-named files.
Observe that files are sorted properly in the list, in a case-sensitive manner.

Review URL: http://codereview.chromium.org/149449

git-svn-id: http://src.chromium.org/svn/trunk/src@20706 4ff67af0-8c30-449e-8e8b-ad334ec8d88c