Fix web contents drag corner case.
We heuristically use the last mouse down event the renderer got as the event that 'started' the drag, which is not entirely accurate. Since the context menu grabs input, we were in a situation where the last mouse down was a right click, so it wanted the drag to end by releasing the right mouse button.

BUG=16793
TEST=see bug

Review URL: http://codereview.chromium.org/155604

git-svn-id: http://src.chromium.org/svn/trunk/src@20825 4ff67af0-8c30-449e-8e8b-ad334ec8d88c